+++ backend-service/cmd/client-grpc/main.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"time"
 
-	v1 "github.com/devzero-inc/grpc-service/backend/pkg/api/service/v1"
+	v1 "github.com/devzero-inc/grpc-service/backend-service/pkg/api/service/v1"
 	"github.com/golang/protobuf/ptypes/empty"
 	"google.golang.org/grpc"
 )
