+++ backend-service/pkg/service/service.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"log"
 
-	v1 "github.com/devzero-inc/grpc-service/backend/pkg/api/service/v1"
+	v1 "github.com/devzero-inc/grpc-service/backend-service/pkg/api/service/v1"
 	"github.com/golang/protobuf/ptypes/empty"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
